make sure physical_memory() groks cgroupsv2

+++ src/login/logind-user.c
@@ -707,7 +707,7 @@ int config_parse_tmpfs_size(
                 void *data,
                 void *userdata) {
 
-        size_t *sz = data;
+        uint64_t *sz = data;
         int r;
 
         assert(filename);
